separate LUT definitions into their own header to be injected into Compile source
